--- .github/workflows/rust-checker.yml
@@ -1,48 +1,0 @@
-name: rust-checker
-
-on:
-  push:
-
-permissions: {}
-
-jobs:
-  run-rustfmt:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@b4ffde65f46336ab88eb53be808477a3936bae11 # v4.1.1
-      - name: Install rust-toolchain
-        uses: dtolnay/rust-toolchain@439cf607258077187679211f12aa6f19af4a0af7 # master
-        with:
-          toolchain: stable
-          components: clippy, rustfmt
-      - name: Run rustfmt using cargo
-        run: cargo fmt --all -- --check
-
-  run-cargo-test:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@b4ffde65f46336ab88eb53be808477a3936bae11 # v4.1.1
-      - name: Install rust-toolchain
-        uses: dtolnay/rust-toolchain@439cf607258077187679211f12aa6f19af4a0af7 # master
-        with:
-          toolchain: stable
-          components: clippy, rustfmt
-
-      - name: Run cargo-test
-        run: cargo test
-
-  run-cargo-clippy:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@b4ffde65f46336ab88eb53be808477a3936bae11 # v4.1.1
-      - name: Install rust-toolchain
-        uses: dtolnay/rust-toolchain@439cf607258077187679211f12aa6f19af4a0af7 # master
-        with:
-          toolchain: stable
-          components: clippy, rustfmt
-
-      - name: Run cargo-clippy -Dwarning
-        run: cargo clippy -- -Dwarnings
