+++ flake.nix
@@ -1,0 +1,61 @@
+{
+  inputs = {
+    nixpkgs.url = "github:nixos/nixpkgs";
+    flake-utils.url = "github:numtide/flake-utils";
+    treefmt-nix.url = "github:numtide/treefmt-nix";
+    rust-overlay.url = "github:oxalica/rust-overlay";
+    crane.url = "github:ipetkov/crane";
+  };
+
+  outputs = { self, nixpkgs, flake-utils, treefmt-nix, rust-overlay, crane }:
+    flake-utils.lib.eachDefaultSystem (system:
+      let
+        pkgs = import nixpkgs { inherit system; overlays = [ (import rust-overlay) ]; };
+        rust = pkgs.rust-bin.fromRustupToolchainFile ./rust-toolchain.toml;
+        treefmtEval = treefmt-nix.lib.evalModule pkgs ./treefmt.nix;
+        craneLib = (crane.mkLib pkgs).overrideToolchain rust;
+        src = ./.;
+        cargoArtifacts = craneLib.buildDepsOnly {
+          inherit src;
+        };
+        kakei = craneLib.buildPackage {
+          inherit src cargoArtifacts;
+          strictDeps = true;
+
+          doCheck = true;
+        };
+        cargo-clippy = craneLib.cargoClippy {
+          inherit src cargoArtifacts;
+          cargoClippyExtraArgs = "--verbose -- --deny warnings";
+        };
+        cargo-doc = craneLib.cargoDoc {
+          inherit src cargoArtifacts;
+        };
+      in
+      {
+        formatter = treefmtEval.config.build.wrapper;
+
+        packages.default = kakei;
+        packages.doc = cargo-doc;
+
+        apps.default = flake-utils.lib.mkApp {
+          drv = self.packages.${system}.default;
+        };
+
+        checks = {
+          inherit kakei cargo-clippy cargo-doc;
+          formatting = treefmtEval.config.build.check self;
+        };
+
+        devShells.default = pkgs.mkShell {
+          packages = [
+            rust
+          ];
+
+          shellHook = ''
+            export PS1="\n[nix-shell:\w]$ "
+          '';
+        };
+      }
+    );
+}
