+++ rust-toolchain.toml
@@ -1,0 +1,4 @@
+[toolchain]
+channel = "1.78.0"
+components = ["rust-src", "rust-analyzer", "llvm-tools-preview"]
+profile = "default"
