+++ src/lib.rs
@@ -1,0 +1,2 @@
+/// contains kakei's sub commands
+pub mod commands;
