+++ treefmt.nix
@@ -1,0 +1,7 @@
+{
+  projectRootFile = "flake.nix";
+  programs.nixpkgs-fmt.enable = true;
+  programs.rustfmt.enable = true;
+  programs.taplo.enable = true;
+  programs.yamlfmt.enable = true;
+}
